refactor(components): extract reusable item card component
The `CardItem` component centralizes the display logic and styling for interactive cards, supporting both internal and external links. This refactoring eliminates duplicated card rendering code in the `/links` and `/` pages, improving consistency and maintainability. A `LinkCardData` interface is also introduced to define the structure for card data.
